Move tests of moved modules
